requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
APScheduler==3.10.4
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
requests==2.31.0
cryptography==41.0.3
web3==6.9.0
eth-account==0.9.0
bitcoin==1.1.42
reportlab==4.0.4
Pillow==10.0.0
python-telegram-bot==20.4
twilio==8.5.0
sendgrid==6.10.0
logging==0.4.9.6
psycopg2-binary==2.9.7
redis==4.6.0
celery==5.3.1
gunicorn==21.2.0